% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{counts_matrix_from_assay}
\alias{counts_matrix_from_assay}
\title{Extract the counts matrix from the Assay}
\usage{
counts_matrix_from_assay(assay)
}
\arguments{
\item{assay}{A SeuratObject::Assay or SeuratObject::Assay5}
}
\value{
A sparse counts matrix
}
\description{
Extract the counts matrix from the Assay
}
