% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{create_bugreport}
\alias{create_bugreport}
\title{Create a Bugreport from a count matrix, projections, and clusters}
\usage{
create_bugreport(
  count_mat,
  clusters,
  projections,
  assay_name = NULL,
  seurat_obj_version = NULL,
  skip_metadata = FALSE
)
}
\arguments{
\item{count_mat}{A sparse dgCMatrix}

\item{clusters}{list of factors that hold information for each barcode}

\item{projections}{list of matrices, all with dimensions (barcodeCount x 2)}

\item{assay_name}{optional string that holds the Seurat Object assay name.}

\item{seurat_obj_version}{optional string that holds the Seurat Object version.  It is useful for debugging compatibility issues.}

\item{skip_metadata}{optional logical which skips printing metadata}
}
\description{
This bugreport can then be included when reaching out to 10xGenomics Support or when filing
a Github ticket.  This information should be included along with any other output when creating a Loupe file.
}
