% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{create_bugreport_from_seurat}
\alias{create_bugreport_from_seurat}
\title{Create a Bugreport from a Seurat Object}
\usage{
create_bugreport_from_seurat(obj)
}
\arguments{
\item{obj}{A Seurat Object}
}
\description{
This bugreport can then be included when reaching out to 10xGenomics Support or when filing
a Github ticket.  This information should be included along with any other output when creating a Loupe file.
}
