% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{create_loupe}
\alias{create_loupe}
\title{Create a Loupe file}
\usage{
create_loupe(
  count_mat,
  clusters = list(),
  projections = list(),
  output_dir = NULL,
  output_name = NULL,
  feature_ids = NULL,
  executable_path = NULL,
  force = FALSE,
  seurat_obj_version = NULL
)
}
\arguments{
\item{count_mat}{A sparse dgCMatrix as is generated via Matrix::rsparsematrix.
Rows are features, Columns are barcodes.}

\item{clusters}{list of factors that hold information for each barcode}

\item{projections}{list of matrices, all with dimensions (barcodeCount x 2)}

\item{output_dir}{optional directory where the Loupe file will be written}

\item{output_name}{optional name of the Loupe file with the extensions not included.}

\item{feature_ids}{optional character vector that specifies the feature ids of the count matrix.
Typically, these are the ensemble ids.}

\item{executable_path}{optional path to the louper executable.}

\item{force}{optional logical as to whether we should overwrite an already existing file}

\item{seurat_obj_version}{optional string that holds the Seurat Object version.
It is useful for debugging compatibility issues.}
}
\value{
TRUE on success, FALSE on error
}
\description{
Create a Loupe file
}
