% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{create_loupe_from_seurat}
\alias{create_loupe_from_seurat}
\title{Create a Loupe file from a Seurat Object}
\usage{
create_loupe_from_seurat(
  obj,
  output_dir = NULL,
  output_name = NULL,
  metadata_cols = NULL,
  dedup_clusters = FALSE,
  feature_ids = NULL,
  executable_path = NULL,
  force = FALSE
)
}
\arguments{
\item{obj}{A Seurat Object}

\item{output_dir}{optional directory where the Loupe file will be written}

\item{output_name}{optional name of the Loupe file with the extensions not included.}

\item{metadata_cols}{optional list that specifies which metadata columns to retain when selecting clusters}

\item{dedup_clusters}{optional logical that will try to deduplicate all clusters that are numerically the same}

\item{feature_ids}{optional character vector that specifies the feature ids of the count matrix.
Typically, these are the ensemble ids.}

\item{executable_path}{optional path to the louper executable.}

\item{force}{optional logical as to whether we should overwrite an already existing file}
}
\value{
TRUE on success, FALSE on error
}
\description{
\code{create_loupe_from_seurat()} passes the active counts matrix,
reductions, and factors found in \code{meta.data} to create a Loupe file.
}
