% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_feature_ids_from_tsv}
\alias{read_feature_ids_from_tsv}
\title{Read FeatureIds from 10x features.tsv.gz file}
\usage{
read_feature_ids_from_tsv(tsv_path)
}
\arguments{
\item{tsv_path}{character vector path to the features.tsv.gz file}
}
\value{
A character vector of the feature ids
}
\description{
Read FeatureIds from 10x features.tsv.gz file
}
