% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{select_assay}
\alias{select_assay}
\title{Select the "best" assay for conversion}
\usage{
select_assay(obj)
}
\arguments{
\item{obj}{A Seurat Object}
}
\value{
A list with the named Seurat Assay or NULL if not found
}
\description{
Prioritizes the active assay, then RNA, and then the rest
Usable assays must have a non empty count matrix
}
