% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{select_projections}
\alias{select_projections}
\title{Select projections from the assay}
\usage{
select_projections(obj)
}
\arguments{
\item{obj}{A Seurat Object}
}
\value{
A list of matrices, all with dimensions (barcodeCount x 2)
}
\description{
Select projections from the assay
}
