% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup eula and download executable}
\usage{
setup(executable_path = NULL)
}
\arguments{
\item{executable_path}{optional string to a non default executable path}
}
\description{
Setup eula and download executable
}
