% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_barcodes}
\alias{validate_barcodes}
\title{Validate the format of the barcodes}
\usage{
validate_barcodes(barcodes)
}
\arguments{
\item{barcodes}{a character vector}
}
\value{
A list with two elements:
\itemize{
\item success: a logical value indicating success (TRUE) or failure (FALSE)
\item msg: an optional error message (NULL if success is TRUE)
}
}
\description{
Validate the format of the barcodes
}
