% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_clusters}
\alias{validate_clusters}
\title{Validate the seurat clusters}
\usage{
validate_clusters(clusters, barcode_count)
}
\arguments{
\item{clusters}{list of factors that hold information for each barcode}

\item{barcode_count}{number of barcodes}
}
\value{
A list with two elements:
\itemize{
\item success: a logical value indicating success (TRUE) or failure (FALSE)
\item msg: an optional error message (NULL if success is TRUE)
}
}
\description{
Validate the seurat clusters
}
