% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_projections}
\alias{validate_projections}
\title{Validate the seurat projections}
\usage{
validate_projections(projections, barcode_count)
}
\arguments{
\item{projections}{list of matrices, all with dimensions (barcodeCount x 2)}

\item{barcode_count}{number of barcodes}
}
\value{
A list with two elements:
\itemize{
\item success: a logical value indicating success (TRUE) or failure (FALSE)
\item msg: an optional error message (NULL if success is TRUE)
}
}
\description{
Validate the seurat projections
}
