% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_count_mat}
\alias{validate_count_mat}
\title{Validate the seurat count matrix}
\usage{
validate_count_mat(count_mat, feature_ids = NULL)
}
\arguments{
\item{count_mat}{A sparse dgCMatrix as is generated via Matrix::rsparsematrix.  Rows are features, Columns are barcodes.}

\item{feature_ids}{optional character vector that specifies the feature ids of the count matrix.  Typically, these are the ensemble ids.}
}
\value{
A list with two elements:
\itemize{
\item success: a logical value indicating success (TRUE) or failure (FALSE)
\item msg: an optional error message (NULL if success is TRUE)
}
}
\description{
Validate the seurat count matrix
}
